+++ src/test/resources/edgeCases/arrayNestedWhere.php
@@ -1,0 +1,9 @@
+<?php
+
+class User extends \Illuminate\Database\Eloquent\Model {
+
+}
+
+User::where([
+    ['email', '=', 'test@email.com'],
+]);